extract skill details side panel from skill details sheet
